Mitigate CSRF attack
Mitigation is relatively easy if you're using an application like Rails.
Many modern frameworks have built in CSRF protection.

To enable CSRF protection in Rails add forgery protection to your
`ApplicationController` to protect sensitive requests. Rails does not
protect GET requests becasuse sensitive requests should never be sent
via GET.

When a request is sent to the application it compares the authenticity
token set in the users session to the token sent in the request. If they
don't match the request will be rejected with an exception. If there is
no token and the application expects a token the request will also be
rejected.