Mitigate XSS in user website
Validating user input before storing it in the database is another way
to protect users from XSS attacks.

This simple validator uses Ruby's built in URI parser to enure that the
scheme provided is `http` or `https`, disallowing the `javascript`
protocol.

Always use a whitelist over a blacklist becasue it default deny protects
the application from accidently allowing a dangerous protocol.

Additionally, use a URI parser when possible over a regex. Regex is easy
to get wrong and URI parsers are widely used and built specifically for
validating URIs and it's harder to trick the URI parser into accepting a
malformed URI than a regex.