Mitigate XXE by disallowing entity replacement
Allowing entity replacement is at the heart of this vulnerability. If
your application is using an XML parsing library that supports entity
replacement it's best to ensure it's off.

If your application is using LibXML you can check if entity replacement
is enabled by running this in the console:

```ruby
LibXML::XML.default_substitute_entities
```

If it returns false, then entity replacement is not supported.